+++ src/XRoadLib/Attributes/XRoadXmlElementAttribute.cs
@@ -19,6 +19,11 @@ namespace XRoadLib.Attributes
         public bool IsOptional { get; set; }
 
         /// <summary>
+        /// Use MTOM/XOP standard for binary content serialization (default value is `true`).
+        /// </summary>
+        public bool UseXop { get; set; } = true;
+
+        /// <summary>
         /// Initializes a new instance of the XRoadXmlElementAttribute class.
         /// </summary>
         public XRoadXmlElementAttribute()